AIBackend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.26.2
tensorflow==2.15.0
gunicorn
scikit-image
scipy
h5py==3.10.0
Werkzeug